the great refactoring
refactored the whole items flags to use the flag system in foundry, this will make adding new flags less painful. also made the character sheet keep the scroll position of the tabs when the data is updated. Also made talents update their flags when their specialisation is updated on the actor sheet.