Stuff of nightmares
Implemented stuff of nightmares and the interaction it has with different weapon attacks. Changed the critical effect crit blurps logic. Changed deathdealer to be less strict. made the crit portion of daemonbane to work. Fixed apply dead when used by a player and not the gm. Fixed chain weapon mastery